requirements.txt
PyYAML==5.4.1
kubernetes==18.20.0
kopf==1.33.0
